Add BulkWrite test and refactor existing tests
Added a new test for BulkWrite functionality in the 'BulkWriterTest.cs' file. Existing tests in 'RepositoryTest.cs' were refactored to inherit common set up from 'RepositoryBaseTest.cs'. An EntityId property was also added in the 'PersonMockupFactory.cs'. The method signature of 'BulkWrite' in 'GenericRepository.cs' was changed to include a 'timeOutInSeconds' parameter.